moved MeetTheTeam and Navbar components to Components folder
