+++ src/Components/MeetTheTeam.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './style/MeetTheTeam.css';
+import '../style/MeetTheTeam.css';
 
 function MeetTheTeam(){
