+++ src/Components/Navbar.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './style/Navbar.css'
+import '../style/Navbar.css'
 
 export const Navbar = () => {
     return (
