+++ src/style/MeetTheTeam.css
@@ -1,71 +1,67 @@
 .meet-the-team-section {
-    font-family: "Poppins", sans-serif;
-    margin-bottom: 100px;
-    margin-top: 100px;  
-  }
-  
-  .meet-the-team-h2 {
-    font-size: 42px;
-    font-weight: 500;
-    line-height: 1.66667;
-    text-align: center;  
-  }
-  
-  .meet-the-team-main-container {
-    display: flex;
-    flex-wrap: wrap;
-  }
-  
-  .meet-team-block {
-    display: flex;
-    margin: auto;
-    position: relative;
-  }
-  
-  
-  .meet-team-block img {
-    filter: grayscale(100%);
-    height: auto;
-    margin: 0;
-  }
-  
-  .meet-team-block img:hover {
-    filter: grayscale(0%);
-  }
-  
-  .meet-team-text-block
-  {
-      background-color: #11111180;
-      bottom: 10px;
-      color: white;
-      padding: 10px;
-      position: absolute;     
-      text-align: center;
-      transition: transform .2s ease-in;    
-      width: 200px;
-      z-index: 1000;
-  }
-  
-  .meet-team-text-block:hover {
-    width: 300px;
-  }
-  
-  .meet-the-team-h4
-  {
-    font-size: 21px;
-    font-weight: 400;
-    margin: 0;  
-  }
-  
-  .meet-the-team-h4:hover {
-    color: #7c73fa;
-    transition: all 0.3s;
-  }
-  
-  .meet-the-team-p
-  {
-    font-weight: 400;
-    letter-spacing: .025em;
-    line-height: 1.5;    
-    margin: 0;
-  }
+  font-family: 'Poppins', sans-serif;
+  margin-bottom: 100px;
+  margin-top: 100px;
+}
+
+.meet-the-team-h2 {
+  font-size: 42px;
+  font-weight: 500;
+  line-height: 1.66667;
+  text-align: center;
+}
+
+.meet-the-team-main-container {
+  display: flex;
+  flex-wrap: wrap;
+}
+
+.meet-team-block {
+  display: flex;
+  margin: auto;
+  position: relative;
+}
+
+.meet-team-block img {
+  filter: grayscale(100%);
+  height: auto;
+  margin: 0;
+}
+
+.meet-team-block img:hover {
+  filter: grayscale(0%);
+}
+
+.meet-team-text-block {
+  background-color: #11111180;
+  bottom: 10px;
+  color: white;
+  padding: 10px;
+  position: absolute;
+  text-align: center;
+  transition: transform 0.2s ease-in;
+  width: 200px;
+  z-index: 20;
+}
+
+.meet-team-text-block:hover {
+  width: 300px;
+}
+
+.meet-the-team-h4 {
+  font-size: 21px;
+  font-weight: 400;
+  margin: 0;
+}
+
+.meet-the-team-h4:hover {
+  color: #7c73fa;
+  transition: all 0.3s;
+}
+
+.meet-the-team-p {
+  font-weight: 400;
+  letter-spacing: 0.025em;
+  line-height: 1.5;
+  margin: 0;
+}
